+++ NEWS.md
@@ -1,9 +1,9 @@
 # ggblanket 2.0.9000
 
 * Breaking: removed gg_function.
+* Breaking: made col_rescale require a function (e.g. scales::rescale()).
 * Improved gg_sf default alpha.
 * Supported gg_sf to work with non-standard named geometry.
-* Fixed gg_raster bug where legend was not showing. 
 * Adjusted gg_theme to work better with ggtext. 
 * Adjusted theme defaults (e.g. title positioning).